create homecontainermixin
