+++ UI/src/components/mixins/RipaHomeContainerMixin.vue
@@ -1,0 +1,83 @@
+<script>
+import { defaultStop, motorStop, probationStop } from '@/utilities/stop'
+import { sampleStop } from '@/stories/data/formStop'
+
+export default {
+  methods: {
+    handleInput(newVal) {
+      this.stop = Object.assign({}, newVal)
+      this.updateFullStop()
+    },
+
+    handleAddPerson() {
+      const updatedStop = this.stop
+      this.stop = Object.assign({}, updatedStop)
+      this.stop.person = {
+        id: new Date().getTime(),
+        isStudent: false,
+        perceivedRace: null,
+        perceivedGender: null,
+        perceivedLgbt: false,
+        perceivedAge: null,
+        anyDisabilities: false,
+        perceivedOrKnownDisability: null,
+      }
+      this.updateFullStop()
+    },
+
+    handleDeletePerson(id) {
+      const filteredPeople = this.fullStop.people.filter(item => item.id !== id)
+      const updatedFullStop = {
+        ...this.fullStop,
+        people: filteredPeople,
+      }
+      this.fullStop = Object.assign({}, updatedFullStop)
+    },
+
+    handleTemplate(value) {
+      this.isEditingForm = true
+
+      if (value === 'motor') {
+        this.stop = motorStop()
+      }
+
+      if (value === 'probation') {
+        this.stop = probationStop()
+      }
+
+      if (value === 'test') {
+        this.stop = sampleStop
+      }
+
+      this.updateFullStop()
+    },
+
+    updateFullStop() {
+      const updatedPerson = {
+        ...this.stop.person,
+        id: this.stop.person.id,
+        actionsTaken: this.stop.actionsTaken,
+        stopReason: this.stop.stopReason,
+        stopResult: this.stop.stopResult,
+      }
+
+      const updatedFullStop = Object.assign({}, this.fullStop)
+      updatedFullStop.updated = new Date()
+      updatedFullStop.officer = this.stop.officer
+      updatedFullStop.stopDate = this.stop.stopDate
+      updatedFullStop.location = this.stop.location
+      const personId = this.stop.person.id
+      const people = updatedFullStop.people || []
+      updatedFullStop.people = people.filter(item => item.id !== personId)
+      updatedFullStop.people.push(updatedPerson)
+      this.fullStop = Object.assign({}, updatedFullStop)
+    },
+
+    handleCancel() {
+      this.isEditingForm = false
+      this.stop = defaultStop()
+      this.updateFullStop()
+    },
+  },
+}
+</script>
